% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_weight.R
\name{count_by_weight}
\alias{count_by_weight}
\title{Count the number of observations, weighted by a weight column}
\usage{
count_by_weight(x, column)
}
\arguments{
\item{x}{The input data. (data.frame).}

\item{column}{Some column in the input data.}
}
\value{
A data.frame or tibble, depending on the input.
}
\description{
Count the number of observations, weighted by a weight column
}
\examples{
df <- data.frame(x = letters[1:3], weight = 1)
count_by_weight(df, "x")

tbl <- tibble::tibble(
  bodyPart = c("head", "head", "tail", "middle"),
  weight = c(10, 11, 8, 30)
)
count_by_weight(tbl, "bodyPart")
}
