% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er.injuries.sandpit-package.R
\docType{package}
\name{er.injuries.sandpit-package}
\alias{er.injuries.sandpit}
\alias{er.injuries.sandpit-package}
\title{er.injuries.sandpit: ER App From "Mastering Shiny"}
\description{
This is a sandbox app for learning about various shiny tooling (testing machinery,
    golem, modules, ...). Hadley Wickham wrote the original source code.
}
\author{
\strong{Maintainer}: Russ Hyde \email{russ.hyde@somewhere.com}

}
\keyword{internal}
