% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{injuries}
\alias{injuries}
\title{Injuries data from NEISS}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 255064 rows and 10 columns.
}
\usage{
injuries
}
\description{
Injuries data from NEISS
}
\keyword{datasets}
