% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population data from NEISS}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 170 rows and 3 columns.
}
\usage{
population
}
\description{
Population data from NEISS
}
\keyword{datasets}
