% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products}
\alias{products}
\title{Products data from NEISS}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 38 rows and 2 columns.
}
\usage{
products
}
\description{
Products data from NEISS
}
\keyword{datasets}
