% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er.injuries.sandpit-package.R
\docType{package}
\name{er.injuries.sandpit-package}
\alias{er.injuries.sandpit}
\alias{er.injuries.sandpit-package}
\title{er.injuries.sandpit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
